Separate destination generation into own file
